requirements.txt
gdown==5.2.0
lmdb==1.4.1
matplotlib==3.8.0
nibabel==4.0.2
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.3
scipy==1.13.1
skimage==0.0
tensorboardX==2.6.2.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.65.0
wandb==0.16.3
